Made more changes to the code almost at testing.

+++ DEA/HelperClasses/FolderFunctions/FolderCleaner.cs
@@ -5,36 +5,38 @@ using Renci.SshNet;
 using WriteLog;
 using WriteNamesToLog;
 
-namespace FolderCleaner;
+namespace DEA.Next.HelperClasses.FolderFunctions;
 
 internal class FolderCleanerClass
 {
     /// <summary>
     /// This function calls the folder cleaning function below.
     /// </summary>
+    /// <param name="downloadFilePath"></param>
     /// <param name="jsonFileNames"></param>
-    /// <param name="customerID"></param>
+    /// <param name="customerId"></param>
     /// <param name="clientEmail"></param>
+    /// <param name="deliveryType"></param>
     /// <returns></returns>
     public static async Task<bool> GetFolders(string downloadFilePath,
         string[] jsonFileNames,
-        int? customerID,
-        string clientEmail,
+        Guid? customerId,
+        string? clientEmail,
         string deliveryType)
     {
-        bool result = false;
-        string localDownloadFilePath = downloadFilePath;
+        var result = false;
+        var localDownloadFilePath = downloadFilePath;
 
         if (deliveryType == MagicWords.Email)
         {
             localDownloadFilePath = Path.GetDirectoryName(downloadFilePath);
         }
 
-        if (Directory.Exists(Path.GetDirectoryName(localDownloadFilePath)))
-        {
-            WriteLogClass.WriteToLog(1, "Cleaning download folder ....", 1);
-            result = await FolderCleaningProcess(localDownloadFilePath, jsonFileNames, customerID, clientEmail, deliveryType);
-        }
+        if (!Directory.Exists(Path.GetDirectoryName(localDownloadFilePath))) return result;
+        WriteLogClass.WriteToLog(1, "Cleaning download folder ....", 1);
+        if (localDownloadFilePath != null)
+            result = await FolderCleaningProcess(localDownloadFilePath, jsonFileNames, customerId, clientEmail,
+                deliveryType);
         return result;
     }
 
@@ -46,16 +48,17 @@ internal class FolderCleanerClass
     /// <param name="jsonFileList"></param>
     /// <param name="customerId"></param>
     /// <param name="clientEmail"></param>
+    /// <param name="deliverType"></param>
     /// <returns></returns>
     private static async Task<bool> FolderCleaningProcess(string downloadedFolderPath,
         string[] jsonFileList,
-        int? customerId,
+        Guid? customerId,
         string clientEmail,
         string deliverType)
     {
         try
         {
-            bool fileMoveResult = true; // Store the result of move files to error folder.
+            var fileMoveResult = true; // Store the result of move files to error folder.
 
             if (deliverType == MagicWords.Email)
             {
@@ -72,35 +75,35 @@ internal class FolderCleanerClass
                 return false;
             }
 
-            IEnumerable<string> nameList = CheckMissedFiles(downloadedFolderPath, jsonFileList);
-            if (nameList.Any()) // If there are any unmatched files.
+            var nameList = CheckMissedFiles(downloadedFolderPath, jsonFileList).ToList();
+            
+            if (nameList.Count != 0) // If there are any unmatched files.
             {
                 // Calls the MoveFilesToErrorFolder method to start moving the missed files.
                 fileMoveResult = await HandleErrorFilesClass.MoveFilesToErrorFolder(downloadedFolderPath,
                     nameList,
                     customerId,
                     clientEmail);
+                
                 // Writes the result to the log.
                 WriteLogClass.WriteToLog(1, fileMoveResult ? $"Moved files {WriteNamesToLogClass.WriteMissedFilenames(nameList)}"
                     : "Moving files was unsuccessful ...", 1);
             }
 
             // Checking if the folder is not empty.
-            IEnumerable<string> fileList = Directory.EnumerateFiles(downloadedFolderPath, "*.*");
+            var fileList = Directory.EnumerateFiles(downloadedFolderPath, "*.*").ToList();
+            
             // Return false if the folder is not empty.
-            if (fileList.Any())
+            if (fileList.Count != 0)
             {
                 return false;
             }
 
-            // Folder delete failes will be written to the log. And return false.
-            if (!DeleteEmptyFolders(downloadedFolderPath))
-            {
-                WriteLogClass.WriteToLog(0, "Deletting empty folders failed ....", 1);
-                return false;
-            }
-
-            return fileMoveResult;
+            // Folder delete files will be written to the log. And return false.
+            if (DeleteEmptyFolders(downloadedFolderPath)) return fileMoveResult;
+            
+            WriteLogClass.WriteToLog(0, "Deleting empty folders failed ....", 1);
+            return false;
         }
         catch (Exception ex)
         {
@@ -113,50 +116,50 @@ internal class FolderCleanerClass
     /// Start the FTP file delete process.
     /// </summary>
     /// <param name="ftpConnect">Ftp connection token.</param>
+    /// <param name="sftpConnect"></param>
     /// <param name="ftpFileList">File list from the FTP server.</param>
     /// <param name="localFileList">File list from the local download folder.</param>
     /// <returns>The result of remove process or false.</returns>
     public static async Task<bool> StartFtpFileDelete(AsyncFtpClient? ftpConnect,
         SftpClient? sftpConnect,
-        IEnumerable<string> ftpFileList,
+        string[] ftpFileList,
         string[] localFileList)
     {
         try
         {
             // Get the FTP path from the first FTP file.
-            string ftpPath = GetFtpPath(ftpFileList.FirstOrDefault());
+            var ftpPath = GetFtpPath(ftpFileList.FirstOrDefault() ?? string.Empty);
             // Local file name list.
-            IEnumerable<string> localFileNames = localFileList.Select(localFilePath => Path.GetFileName(localFilePath));
+            var localFileNames = localFileList.Select(Path.GetFileName);
             // FTP file name list.
-            IEnumerable<string> ftpFileNames = ftpFileList.Select(ftpFilePath => Path.GetFileName(ftpFilePath));
+            var ftpFileNames = ftpFileList.Select(Path.GetFileName);
             // Matching file names.
-            IEnumerable<string> matchingFileNames = localFileNames.Intersect(ftpFileNames);
+            var matchingFileNames = localFileNames.Intersect(ftpFileNames).ToList();
             // Result of the foreach loop.
-            bool result = false;
+            var result = false;
 
             // If there are no matching files, return.
-            if (!matchingFileNames.Any())
+            if (matchingFileNames.Count == 0)
             {
                 return result;
             }
 
-            if (ftpConnect != null)
+            if (ftpConnect == null) return result;
+            
+            // Delete FTP files.
+            foreach (var matchingFileName in matchingFileNames)
             {
-                // Delete FTP files.
-                foreach (string matchingFileName in matchingFileNames)
-                {
-                    result = await DeleteFtpFiles(ftpConnect, string.Concat(ftpPath, matchingFileName));
-                }
+                result = await DeleteFtpFiles(ftpConnect, string.Concat(ftpPath, matchingFileName));
             }
 
-            if (sftpConnect != null)
+            if (sftpConnect == null) return result;
+            
+            // Delete SFTP files.
+            foreach (var matchingFileName in matchingFileNames)
             {
-                // Delete SFTP files.
-                foreach (string matchingFileName in matchingFileNames)
-                {
-                    result = await DeleteSftpFiles(sftpConnect, string.Concat(ftpPath, matchingFileName));
-                }
+                result = await DeleteSftpFiles(sftpConnect, string.Concat(ftpPath, matchingFileName));
             }
+            
             return result;
         }
         catch (Exception ex)
@@ -172,29 +175,29 @@ internal class FolderCleanerClass
     /// <param name="localFolderPath">Local download process path</param>
     /// <param name="remoteFileList">FTP files list.</param>
     /// <returns>Returns the unmatched file names list.</returns>
-    private static IEnumerable<string> CheckMissedFiles(string localFolderPath,
+    private static List<string> CheckMissedFiles(string localFolderPath,
         IEnumerable<string> remoteFileList)
     {
         try
         {
             // Makes the downloaded files list from the folder path.
-            IEnumerable<string> downloadedFileList = Directory.EnumerateFiles(localFolderPath, "*.*");
+            var downloadedFileList = Directory.EnumerateFiles(localFolderPath, "*.*");
             // Creates file name only list from the json file list.
-            IEnumerable<string> jsonFileNames = remoteFileList.Select(jsonFilePath => Path.GetFileName(jsonFilePath));
+            var jsonFileNames = remoteFileList.Select(jsonFilePath => Path.GetFileName(jsonFilePath));
             // Creates file name only list from the downloaded file list.
-            IEnumerable<string> downloadedFileNames = downloadedFileList.Select(downloadedFilePath => Path.GetFileName(downloadedFilePath));
+            var downloadedFileNames = downloadedFileList.Select(downloadedFilePath => Path.GetFileName(downloadedFilePath));
             // Gets the unmatched file names. From matching the above two lists.
             //IEnumerable<string> unmatchedFileNames = jsonFileNames.Except(downloadedFileNames).Concat(downloadedFileNames.Except(jsonFileNames));
-            IEnumerable<string> matchedFileNames = downloadedFileNames.Intersect(jsonFileNames);
+            var matchedFileNames = downloadedFileNames.Intersect(jsonFileNames).ToList();
 
-            WriteLogClass.WriteToLog(1, !matchedFileNames.Any() ? "No missed files found ...."
-                : $"Found {matchedFileNames.Count()} missed files ....", 1);
+            WriteLogClass.WriteToLog(1, matchedFileNames.Count == 0 ? "No missed files found ...."
+                : $"Found {matchedFileNames.Count} missed files ....", 1);
             return matchedFileNames;
         }
         catch (Exception ex)
         {
             WriteLogClass.WriteToLog(0, $"Exception at check missed files: {ex.Message}", 0);
-            return Enumerable.Empty<string>();
+            return [];
         }
     }
 
@@ -202,7 +205,7 @@ internal class FolderCleanerClass
     /// Deletes files from the FTP server.
     /// </summary>
     /// <param name="ftpConnect">FTP connection token.</param>
-    /// <param name="ftpFileName">FTP files list.</param>
+    /// <param name="fileToDelete"></param>
     /// <returns>Return true or false.</returns>
     private static async Task<bool> DeleteFtpFiles(AsyncFtpClient ftpConnect,
         string fileToDelete)
@@ -254,6 +257,7 @@ internal class FolderCleanerClass
     /// Delete all files from the download folder.
     /// </summary>
     /// <param name="downloadFolderPath">Path of the local download folder.</param>
+    /// <param name="jsonFileList"></param>
     /// <returns>Returns true or false.</returns>
     public static bool DeleteFiles(string downloadFolderPath,
         string[] jsonFileList)
@@ -265,25 +269,25 @@ internal class FolderCleanerClass
         }
 
         // Makes the downloaded files list from the folder path.
-        IEnumerable<string> downloadedFileNamesList = Directory.EnumerateFiles(downloadFolderPath, "*.*");
+        var downloadedFileNamesList = Directory.EnumerateFiles(downloadFolderPath, "*.*");
         // Creates file name only list from the downloaded file list.
-        IEnumerable<string> downloadedFileNames = downloadedFileNamesList.Select(downloadedFilePath => Path.GetFileName(downloadedFilePath));
+        var downloadedFileNames = downloadedFileNamesList.Select(Path.GetFileName).ToList();
         // Checking if the jsonFileList contains file names or file names with path.
-        bool containsPath = jsonFileList.Any(filePath => filePath.Contains('/') || filePath.Contains('\\'));
+        var containsPath = jsonFileList.Any(filePath => filePath.Contains('/') || filePath.Contains('\\'));
         // Creates file name only list from the json file list.
-        IEnumerable<string> jsonFileNames = containsPath ? jsonFileList.Select(jsonFilePath => Path.GetFileName(jsonFilePath)) : jsonFileList;
+        var jsonFileNames = containsPath ? jsonFileList.Select(jsonFilePath => Path.GetFileName(jsonFilePath)) : jsonFileList;
         // Makes the matching file names list.
         //IEnumerable<string> matchingFileNames = jsonFileList.Intersect(jsonFileList); <-- remove this if not needed. Do this next release
-        IEnumerable<string> matchingFileNames = jsonFileNames.Intersect(downloadedFileNames).Where(fileName => downloadedFileNames.Contains(fileName));
+        var matchingFileNames = jsonFileNames.Intersect(downloadedFileNames).Where(fileName => downloadedFileNames.Contains(fileName));
         // Result of the foreach loop.
-        bool result = true;
+        var result = true;
 
         // Delete files.
-        foreach (string fileName in matchingFileNames)
+        foreach (var fileName in matchingFileNames)
         {
             try
             {
-                File.Delete(Path.Combine(downloadFolderPath, fileName));
+                if (fileName != null) File.Delete(Path.Combine(downloadFolderPath, fileName));
             }
             catch (Exception ex)
             {
@@ -304,26 +308,28 @@ internal class FolderCleanerClass
     {
         try
         {
-            bool result = true;
+            var result = true;
             // Get the download folder name.
-            string basePath = Path.GetDirectoryName(downloadFolderPath);
+            var basePath = Path.GetDirectoryName(downloadFolderPath);
+            if (basePath == null) return result;
+            
             // List all the folder with in the local download folder.
-            IEnumerable<string> directoryList = Directory.EnumerateDirectories(basePath, "*", SearchOption.AllDirectories);
+            var directoryList = Directory.EnumerateDirectories(basePath, "*", SearchOption.AllDirectories);
             // List all the empty folders.
-            IEnumerable<string> emptyFolderList = directoryList.Where(dirPath => !Directory.EnumerateFileSystemEntries(dirPath).Any());
+            var emptyFolderList = directoryList.Where(dirPath => !Directory.EnumerateFileSystemEntries(dirPath).Any()).ToList();
 
-            if (!emptyFolderList.Any())
+            if (emptyFolderList.Count == 0)
             {
                 WriteLogClass.WriteToLog(1, "No empty folders found ....", 1);
                 return false;
             }
 
-            foreach (string emptyFolder in emptyFolderList)
+            foreach (var emptyFolder in emptyFolderList)
             {
                 try
                 {
                     // Check if the folder is not empty.
-                    IEnumerable<string> fileList = Directory.EnumerateFiles(emptyFolder, "*.*");
+                    var fileList = Directory.EnumerateFiles(emptyFolder, "*.*");
                     if (fileList.Any())
                     {
                         WriteLogClass.WriteToLog(1, "The download folder is not empty ....", 1);
@@ -340,6 +346,7 @@ internal class FolderCleanerClass
                     result = false;
                 }
             }
+
             return result;
         }
         catch (Exception ex)
@@ -349,15 +356,15 @@ internal class FolderCleanerClass
         }
     }
 
-    private static bool AttachmentFileDelete(string downloadFolerPath,
+    private static bool AttachmentFileDelete(string downloadFolderPath,
         string[] jsonFileList)
     {
-        bool result = true;
+        var result = true;
         try
         {
-            foreach (string jsonFile in jsonFileList)
+            foreach (var jsonFile in jsonFileList)
             {
-                string fullFileName = Path.Combine(downloadFolerPath, jsonFile);
+                var fullFileName = Path.Combine(downloadFolderPath, jsonFile);
                 File.Delete(fullFileName);
             }
         }
